+++ CHANGELOG.md
@@ -23,6 +23,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
  - Info command
  - Function to make API calls to Torn
  - Additional commands in the Torn cog (e.g. add key)
+ - Discord User ID check in vault commands
 
 ## Changed
 - Withdrawal command to have a specific on reaction loop instead of a general loop